Bug 28316: escape brackets in the search query
This patch screens square and curly brackets which have no special
language meaning, while keeping ranges as they are, allowing ES to
search for biblios with titles that contain square and curly brackets
and at the same time allows the use of range searches ([1990 TO 2000]).

To reproduce:
1) using ES, search for the book with title that contains
square and/or curly brackets, like "book [second edition]", which will
result in error.
2) apply the patch.
3) search for that book again, ensure that it works now.
4) additionally perform a search using range,
for example: "[1999 TO 2020]", ensure that it works correctly too.