Bug 28736: (QA follow-up) hints in explanation about broken syntax
Add direct hint for the user what might be the issues
with search phrase.

In most cases we have unescaped or not removed special characters
when user copy-pastes book name. If we explain that error happens
because there is special search language, user will know what to
research because we explicitly will tell that this is NOT the
empty result, but search wasn't performed.