Bug 00000: HOTFIX. Bug: empty date becomes empty list and breaks a hash.
dt_from_string with date as '0000-00-00 00:00:00' does just `return`,
but just `return` in the list context means `()` (empty list), that
lead to warnings (and holes) like this:

Odd number of elements in anonymous hash at .../Koha/Object.pm line 430.

in:
    dt => dt_from_string($unblessed->{$col}, 'sql'),

if we will `return undef` we explicitly will prevent return of the empty
list in the list contexts.