Bug 33568: Hide empty columns
Instead of relying on the existence of data in at least one item we are
going to hide empty columns.
That could be confusing: it does not mean items on another page do not
have data for this column. The user can still display the column using
the "Columns" visibility button, and filter on the column.

I think this is an acceptable behaviour.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>