Bug 38666: Closed stack requests
A closed stack request is a special kind of hold that can only be placed
on a closed stack item.
A closed stack item is an item that is not available publicly in the
library but can be made available on demand. When a request is made, a
librarian will print a slip then will go retrieve the document and place
the slip in place of the document.

Internally, a closed stack item is an item that is flagged as "closed
stack" (new column items.is_closed_stack). The flag can be modified on
the biblio "Items" tab. The new column can also be mapped to an item
MARC subfield.

An item is available for a closed stack request only if:
- it is flagged as "closed stack"
- there is no holds on it
- it is not checked out
- it is not in transfer

If any of the above conditions is false, it is not available for a
closed stack request but it can be available for a regular hold.
In particular this means that once a closed stack request has been made
on a particular item, this item can now be reserved using regular holds.

Closed stack requests behave like regular holds, with the following
exceptions:
- They can only be placed on a specific item (item-level hold is forced)
- They can be placed on "not for loan" items
- They bypass the "onshelfholds" circulation rule
- They cannot be cancelled from OPAC

There is a new page "Closed stack requests" accessible from the
circulation menu that lists all pending closed stack requests.
From there it is possible to cancel requests or to print a slip (letter
code CLOSED_STACK_SLIP in 'reserves' module)
Once the slip is printed, closed stack requests move to another tab in
the same page.

The patch also adds a template plugin 'GD.Barcode' that allows to
include a barcode image inside a letter. Useful to be able to scan the
printed closed stack request slip.
Example usage:

[% USE GD.Barcode %]
<img src="[% GD.Barcode.create_as_data_url('Code39', hold.item.barcode) | html %]">

Test plan:
 1. Apply the patch:
    a. apply patch,
    b. run updatedatabase.pl and `update_dbix_class_files.pl --force`,
    c. run `yarn run build`,
    d. and finally restart koha
 2. Create a biblio with two items.
 3. On the biblio detail page, click on the "Items" tab
 4. For one of the two items, change the "Closed stack" status from No
    to Yes
 5. Notice there is now a "closed stack requests" tab on the left, and a
    "closed stack request" button in the toolbar
 6. Click on "Place hold" and select a patron. Notice there is only one
    item displayed in the table (the non-"closed stack" item)
 7. Go back to the biblio detail view and click on the "Closed stack
    request" button. Select a patron. Notice there is only one item
    displayed in the table (the "closed stack" item).
 8. Select this item and click on "Place hold". You should be redirected
    to the "Closed stack requests" tab where you can see the request
    that you just made
 9. Try to place a regular hold again. This time you should see both
   items in the table.
10. Go to Tools » Notices and slips
11. Create a new notice in the Holds module with code CLOSED_STACK_SLIP.
    Add content in the "Print" message.
12. Go to Circulation » Closed stack requests
13. You have two tabs: "Pending" and "Slip printed", you should have one
    request under the pending tab, and zero under the other one.
14. Click on the "Print closed stack request slip" button, in the
    "pending" table. A printer dialog should open, close it.
    The "Closed stack requests" should have been automatically
    refreshed, showing you an empty "pending" tab, while the "Slip
    printed" tab now contains the previously pending request.
15. Cancel the closed stack request.
16. Login to OPAC, find your biblio record and place a closed stack
    request. Like from the staff interface, only the closed stack item
    should be in the items table, and it should not be possible to place
    a biblio-level request
17. In your account page, verify that the closed stack request appear
    under the new tab "Closed stack requests" and that there is no
    "Cancel" button
18. Go back to the staff interface main page. You should see an alert at
    the bottom: "Pending closed stack requests: 1". Clicking on it
    should redirect you to the closed stack requests page seen earlier
    (step 12)
19. Set the closed stack item as "notforloan", either by editing the
    item or its item type. Verify that you can still do closed stack
    requests
20. Change circulation rules to forbid on-shelf holds on the closed
    stack item and verify that you can still do closed stack requests.

Sponsored-by: Université de Lyon 3
Signed-off-by: thibault <thibault.keromnes@univ-paris8.fr>