Bug 20447: MARC Holdings support
Enable by setting "SummaryHoldings" option to "Use".

Smaller notes to task parts (batch rebuild script):
---

Includes misc/batchRebuildHoldingsTables.pl: a batch rebuild script for
holdings table which can be used to update the columns in the holdings
table from the MARC record using the rules in the framework. In
practice this will only be required if the koha link field is changed
in the framework.

---
End of smaller notes

Sponsored-by: The National Library of Finland

Co-authored-by: Joonas Kylmälä <joonas.kylmala@helsinki.fi>
Co-authored-by: Andrew Nugged <nugged@gmail.com>

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>