Bug 39770: Fix wrong Braille format search for Elasticsearch
ElasticSearch couldn't find any books that have Braille: No results match your search  with limit(s): 'ff7-01-02:("tc or fb")'. It was a problem because the "l-format:tc or fb" only works for Zebra and is incorrect way to search with OR for ElasticSearch.

Elastic requires the syntax to be l-format:"tc" OR "fb" (uppercase OR and each format in double quotes) instead.

This patch fixes it by checking which search engine you use depending on the current syspref SearchEngine and provides the correctly formatted format.

To reproduce:
1. Have ElasticSearch enabled and check that you have books with braille, or make up one yourself by adding "tc" (text - Braille) or "fb" (tactile material - braille) to 007 MARC field.
2. Go to catalogue/advanced search and select format: braille, from the dropdown.
3. Try to find anything, if you are using ElasticSearch you will see: No results match your search  with limit(s): 'ff7-01-02:("tc or fb")'
4. Do the same with Zebra instead of Elastic, it should work properly and you should see all the books that have Braille.
5. Apply the patch.
6. Repeat same search both using ElasticSearch and then Zebra. Both ElasticSearch and Zebra should work without any issues.