Bug 41338: Hold found notice should show item's home library and where it is currently checked in
This patch adds a description of item's home library and the library where the item is currently
checked in at, to the hold found notice when you check in the item. Also to avoid redundancy,
the home library will be shown only when destination branch is different from the home branch.

To reproduce:
1. Configure three libraries, A, B and C.
2. Create an item with the home library A.
3. Check in the item at the library B.
4. Place a hold with a pickup branch C and then check thought the interface for the same library C.
5. The only information about branch related to the item will be "Transfer to: Fairfield".
6. Click ignore and cancel the transfer.
7. Apply the patch.
8. Check in the item again.
9. This time you should see the full information about home library and where the item is currently checked in at.